Cloud Computing Proiect/Backend/requirements.txt
Flask==1.1.2
SQLAlchemy==1.3.17
pg8000==1.13.2
Flask-API==1.1
google-cloud-storage==1.28.1
gunicorn==20.0.4; python_version > '3.0'
gunicorn==19.10.0; python_version < '3.0'
google-oauth2-tool==0.0.3
Flask-Cors==3.0.8
Werkzeug==1.0.1